+++ project-euler/src/main/scala/raymitchell/euler/util/PrimeNumbers.scala
@@ -3,18 +3,18 @@ package raymitchell.euler.util
 /**
   * Iterable through all prime numbers.
   */
-object PrimeNumbers extends Iterable[Int] {
+object PrimeNumbers extends Iterable[Long] {
   // Cache of previously computed primes
   // TODO: Make thread safe
-  private var primes = Vector[Int]()
+  private var primes = Vector[Long]()
 
-  override def iterator: Iterator[Int] = new Iterator[Int] {
+  override def iterator: Iterator[Long] = new Iterator[Long] {
 
     override def hasNext: Boolean = true
 
-    override def next(): Int = {
+    override def next(): Long = {
       // A number is prime if it has no prime factors
-      def isPrime(x: Int): Boolean = {
+      def isPrime(x: Long): Boolean = {
         primes.forall(x % _ != 0)
       }
